r21683: Remove the gconf registry backend.
It's already been disabled for a while, as it required some compile flags that broke other things.

While this backend was an interesting proof-of-concept, this module doesn't really add any value.
It's not possible to map between windows registry and gconf semantics and I can't think of
any reason why you would want to do so.
(This used to be commit 02481df3e8653ab366cb73d124e52bc9551bb717)