r21879: Move process_blocking_lock_queue to a timed event.
The idea is that we have blocking.c:brl_timeout as a timed
event that is present whenever we do have a blocking lock
pending. It fires brl_timeout_fn() which calls
process_blocking_lock_queue().

Whenever we make changes to blocking_lock_queue, we trigger
a recalc_brl_timeout() which sets a new brl_timout event if 
necessary. This makes the call to
blocking_locks_timeout_ms() in setup_select_timeout() 
unnecessary, this is implicitly done in
event_add_to_select_args() from the timed events.


Volker