r21942: Hoist by our own petard :-). Older smbclient binaries were not able to connect to the rewritten dfs code as they set the dfs flag bit but then send local paths.
Now that our dfs code is a *lot* more robust in
detecting this sort of braindamage we can just
call into it directly on getting a DFS flag
and let the parser sort it out without having
to check it's actually connecting to a dfs
enabled share (I'm proud of this code :-).

Jeremy.
(This used to be commit 8c4d929c76ba56d59f651c57d1feb37d2916a436)