r22160: Only use test environment names that actually exist (get make quicktest to pass).
Ensure that when we fail on the build farm, this ends up in the total
failures.

I think having the 2 different functions for running the tests is just
asking for this kind of failure.  The build farm should be a
presentation detail, and not alter the test behaviour.

There is still a gremlin in there, we will fail horribly on the farm,
but not on local tests...

Andrew Bartlett