r22318: If we're running as non root we might not be able to dump the core file to the corepath.
Even the chdir() will fail if the LOGFILEBASE path is set 0700.

If the currrent user doesn't have the permission to create the core file
we end with:
   unable to change to <LOGFILEBASE>
   refusing to dump core

The alternative would be to change the permissions of the directory.
But taht would not ensure core dumps are working out of the box.