avoid a deadlock the fetch_lock code. The deadlock could happen when a client held the chainlock, and the daemon received a dmaster reply at the same time. The daemon would not be able to process the dmaster reply, due to the lock, but the fetch lock cannot make progres until the dmaster reply is processed.
The solution is to not hold the lock in the client while talking to
the daemon. The client has to retry the lock after the record has
migrated. This means that forward progress is not guaranteed. We'll
have to see if that matters in practice.

(This used to be ctdb commit 737e5a1253cb048222c595a474aff71c99fc554f)