r22401: Add bzr revision+nick support to mkversion.sh.
The plan is to actually synchronize with SAMBA_4_0's mkversion.sh,
which is more elaborate. But there are a few differences in
behaviour I still have to check. So I am cheching this here in for
a start...

Michael
(This used to be commit dec7e31f298c565b49b7bd650e164274d0f61778)