r22705: Implement new set_dc_type_and_flags() called based on the information return from our DC in the DsEnumerateDomainTrusts() call.   If the fails, we callback ot the older connect-to-the-remote-domain method.
Note that this means we can only reliably expect the native_mode
flag to be set for our own domain as this information in not
available outside our primary domain from the trusted information.
This is ok as we only really need the flag when trying to 
determine to enumerate domain local groups via RPC.

Use the AD flag rather than the native_mode flag when using
ldap to obtain the seq_num for a domain.