r22949: Merge modified and extended winbindd cache validation code from 3_0. (Essentially r22845 and r22847).
Original log: (r22845)

Modified and extended the winbindd cache validation code:                             
                                                                                      
* Replaced signal catching/longjmp magic by a fork:                                   
  Let the child do the actual validation of the entries.                              
  Exit code and signals are intercepted by waitpid.                                   
* Fix logic so that also encounter of an unknown key in the                           
  tdb leads to an error.                                                              
* Extended status of validation is kept in a (as yet simple)
  stuct and communicated over a pipe from child to parent.
* Added two validation_ functions for two new keys.

The call of winbindd_validate_cache is still commented out
in the winbindd main loop. But I am currently testing it
and so far it seems to work fine.

The next step in my plan is to generalize the validation
mechanism to a tdb_open_log_validate function in lib/util_tdb.c.
There ist nothing very special about the cache tdb here,
and this might be useful elsewhere...

Michael