r23024: Ok, neither the duplicates_allowed nor the timeout argument to message_send_pid is used anymore. Two users of duplicates_allowed: winbind and the printer notify system.
I don't thing this really changes semantics: duplicates_allowed is hell racy
anyway, we can't guarantee that we don't send the same message in sequence
twice, and I think the only thing we can harm with the print notify is
performance.

For winbind I talked to Günther, and he did not seem too worried.

Volker