r23133: I felt pity on Kai, as he starts work on winbind in Samba4, so I decided to clean it up a little.
We now use SPNEGO for authentication if possible, and common routines
shared with the rest of the librpc codebase.  Rather than make a
connection to IPC$, then connect the pipes to it, we instead have the
lsa and samr pipes as 'secondary connections'.

Andrew Bartlett
(This used to be commit 86654056b22245a57396544d572de6401069b9e5)