Add an initial amount of documentation of registry based configuration to smb.conf(5). This explains registry shares and registry global options and mentions "net conf".
This will be in 3.0.26, but Jerry told me to commit this
anyways. The plan is to branch the doc for 3.0.25b from r1126 (say).

I could not check the result since I could not get the
manpages built correctly on any system I tried. So the
output needs to be checked - sorry, I will fix my
environment later...

Michael
(This used to be commit 2195dc476f4946ff062b62cf29d1b330263507bc)