r23750: Change the behaviour of net conf import when there is a global section in the current registry and there is no global section in the input file (or only global options with default values): In that case the existing global section is now not touched. Before, it would have been deleted and recreated empty. The new behaviour is how other shares are treated too.
Note that since the input file is parsed by lp_load, there is currently
no way to distinguish between a section with only default parameters
and a non-existing section in net conf import.

Michael

PS: A couple of trailing white-spaces have been eliminated
and a line was broken to be not longer than 80 chars, too.