r24286: (missed from previous commit)
A number of machines on the build farm fail because while they use
SOCKET_WRAPPER, they don't compile with --enable-developer.  As such
the winbind tests fail.

We need to ensure that on production systems we never have
SOCKET_WRAPPER set, or this could be a security issue.

Andrew Bartlett
(This used to be commit 3d2dbcf50dbe3d36ec86e1bc1d7bdcb72dbdef00)