r24347: fake_sendfile does not need Inbuf/Outbuf
In the future, we might put the new Linux splice(2) syscall here. This
should also work for reply_write, but getting that in is a bit trickier.
We need to decide very early before fetching the whole buffer that we
have a write call.
(This used to be commit 32921c878a7f60251a9217173951065784077077)