r24916: (merge to release branch)
In response to bug #4892 by Matthias Wallnöfer <mwallnoefer@yahoo.de>
allow the objectclass module to reconstruct the objectclass hierarchy,
rather than using templates.

The issue being fixed in particular is that 'top' was not being set on
containers.

This should ensure we do this right for all objects.

Andrew Bartlett